requirements.txt

pymongo
flask-pymongo
python-dotenv
openai
Flask==2.2.2
gunicorn
Werkzeug==2.2.2
